portlist, util/winutil: change Windows portlist implementation to disambiguate svchost processes.
We query the Windows Service Control Manager for information about service
processes so that we can map services hosted by `svchost.exe` to the real name
of the service being hosted. Unfortunately we need to query the process names of
all running services to determine which ones are `svchost` processes, but we
also temporarily cache them, which should help avoid some redundant system calls.
I think this extra work is worth it, as so many system services are hosted in
`svchost` that we would otherwise lose a lot of information compared to the
previous implementation that shelled out to `netstat`.

This gives us similar fidelity to what we were getting when we were shelling out
to `netstat`, with the added bonus that this implementation appears to be able
to produce slightly better quality data.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>